requirements.txt
easyocr==1.4.1
opencv-python==4.5.2.52
flask~=2.0.2
wfastcgi
flask_restful
pdf2image~=1.16.0
requests~=2.26.0
numpy~=1.21.5
imageio~=2.9.0
Pillow~=8.2.0
tifffile~=2021.11.2